feat: export bind/unbind methods, and companion mixin
There's a particular situation where directives are not the best alternative: Renderless components that exposes their internal functionality via slot scope bindings. Consumers of this componentes usually don't need to be aware of what is binded, just need to do v-bind="slotScope" to their slot component and get on with their lives.

The problem is that we can't pass a directive up via scope. A common counterpoint is "just import it on the consumer and apply to the slot-scope element". This is mostly true, but it's also equivalent to say to to mcdonals customer "here's your big mac! Don't forget to add your burger in the middle before eating, since it's served on the side". — weird right?

- So this commit export of the "public" directive methods (bind, unbind) that would allow use to programatically use v-click-outside package. A companion namespaced mixin can shortcut the directive bindingValue object format, but I understand that mixins are not that popular and vue3 will make them obsolete, so it's totally optional to remove it from this commit.
- updates "examples" with 2 boxes using this version.